Refactor AWSDynamoDBSource's adapter to use eventing lib
The test adapter library does not yet support sending events to a mock
server, so those tests had to be simplified.
