+++ cmd/awsdynamodbsource/kodata/refs
@@ -1,0 +1,1 @@
+../../../.git/refs
