+++ config/500-controller.yaml
@@ -60,6 +60,8 @@ spec:
         # Source adapters
         - name: AWSCODECOMMITSOURCE_IMAGE
           value: ko://github.com/triggermesh/aws-event-sources/cmd/awscodecommitsource
+        - name: AWSDYNAMODBSOURCE_IMAGE
+          value: ko://github.com/triggermesh/aws-event-sources/cmd/awsdynamodbsource
 
         securityContext:
           allowPrivilegeEscalation: false
