+++ pkg/adapter/awsdynamodbsource/events.go
@@ -1,0 +1,32 @@
+/*
+Copyright (c) 2020 TriggerMesh Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package awsdynamodbsource
+
+import (
+	"github.com/aws/aws-sdk-go/service/dynamodbstreams"
+)
+
+// DynamoDBEvent represents an AWS Dynamo DB payload.
+type DynamoDBEvent struct {
+	AwsRegion    *string                       `locationName:"awsRegion" type:"string"`
+	Dynamodb     *dynamodbstreams.StreamRecord `locationName:"dynamodb" type:"structure"`
+	EventID      *string                       `locationName:"eventID" type:"string"`
+	EventName    *string                       `locationName:"eventName" type:"string" enum:"OperationType"`
+	EventSource  *string                       `locationName:"eventSource" type:"string"`
+	EventVersion *string                       `locationName:"eventVersion" type:"string"`
+	UserIdentity *dynamodbstreams.Identity     `locationName:"userIdentity" type:"structure"`
+}
